fix(monitor): attach live-scroll probe to ScrollView itself
The audit captured only 4 perf.dashboard-live-scroll events — all of
them coming from the driver. The view-side trigger and offset events
were missing because `.onScrollGeometryChange` was applied to the
HarnessMonitorColumnScrollView wrapper, which is OUTSIDE the underlying
ScrollView. SwiftUI's geometry-change modifier needs to sit on the
ScrollView itself, so move the probe into the column-scroll-view body
and gate it on whether a perf scroll-position binding is present.

Also add an onAppear trigger event from the dashboard route view so we
can prove the view mounted and the perf hook is active.
