+++ apps/harness-monitor-macos/Sources/HarnessMonitorUIPreviewable/Views/Dashboard/HarnessMonitorPerfDashboardScrollBus.swift
@@ -1,6 +1,5 @@
 import Foundation
 import OSLog
-import SwiftUI
 
 /// Perf-only bus used by `HarnessMonitorPerfDriver` to ask the live dashboard window
 /// to programmatically scroll. The dashboard view listens for these notifications
@@ -56,7 +55,7 @@ public enum HarnessMonitorPerfDashboardScrollBus {
     )
   }
 
-  /// Emit an offset signpost. Called from the dashboard route view's
+  /// Emit an offset signpost. Called from the ScrollView's
   /// `.onScrollGeometryChange` so the trace records every time the actual scroll
   /// position moves. An audit can compare trigger events against offset events to
   /// confirm the scroll surface really moved.
@@ -65,25 +64,5 @@ public enum HarnessMonitorPerfDashboardScrollBus {
       "perf_event",
       "component=perf.dashboard-live-scroll event=scroll.offset y=\(Int(y.rounded()))"
     )
-  }
-}
-
-/// Observes the closest scroll view's content offset and emits a perf signpost on
-/// every change. Only the dashboard live-scroll scenario opts in via `enabled`, so
-/// the modifier stays inert for shipping builds and non-perf surfaces.
-struct DashboardPerfScrollGeometryProbe: ViewModifier {
-  let enabled: Bool
-
-  @ViewBuilder
-  func body(content: Content) -> some View {
-    if enabled {
-      content.onScrollGeometryChange(for: CGFloat.self) { geometry in
-        geometry.contentOffset.y
-      } action: { _, newY in
-        HarnessMonitorPerfDashboardScrollBus.recordOffset(newY)
-      }
-    } else {
-      content
-    }
   }
 }
